Backend/requirements.txt
fastapi
uvicorn
pydantic
motor
python-dotenv
bcrypt
passlib
python-jose
pyjwt
requests
google-cloud-storage